Add files via upload
This is a framework which takes a file of data stored as a python dictionary by year and creates a family tree structure for Big-Little relationships which can support co-bigging, twins, children older than their ancestors, and other weird relationships that hopefully only occur in non-biological families. The interface runs the framework and allows the user to look up an individual and learn their family history and other data. These are sample files, but this framework could be extended to any organization's Big-Little family tree, including those too complex to draw on paper.